mbed validatorAndCollector Refactor
removed unnecessary workers and visitors
simValidators now check the configuration
