--- RobotMbed/src/main/java/de/fhg/iais/roberta/visitor/validate/Calliope2017NoBlueValidatorAndCollectorVisitor.java
@@ -1,13 +1,0 @@
-package de.fhg.iais.roberta.visitor.validate;
-
-import com.google.common.collect.ClassToInstanceMap;
-
-import de.fhg.iais.roberta.bean.IProjectBean;
-import de.fhg.iais.roberta.components.ConfigurationAst;
-
-public final class Calliope2017NoBlueValidatorAndCollectorVisitor extends CalliopeValidatorAndCollectorVisitor {
-
-    public Calliope2017NoBlueValidatorAndCollectorVisitor(ConfigurationAst brickConfiguration, ClassToInstanceMap<IProjectBean.IBuilder> beanBuilders) {
-        super(brickConfiguration, beanBuilders);
-    }
-}
