+++ RobotMbed/src/main/java/de/fhg/iais/roberta/visitor/validate/Calliope2017ValidatorAndCollectorVisitor.java
@@ -1,7 +1,6 @@
 package de.fhg.iais.roberta.visitor.validate;
 
 import com.google.common.collect.ClassToInstanceMap;
-
 import de.fhg.iais.roberta.bean.IProjectBean;
 import de.fhg.iais.roberta.components.ConfigurationAst;
 import de.fhg.iais.roberta.syntax.action.mbed.RadioReceiveAction;
@@ -12,7 +11,7 @@ import de.fhg.iais.roberta.syntax.sensor.mbed.RadioRssiSensor;
 /**
  * Calliope 2017 has Bluetooth Support => The radio block must be disabled!
  */
-public final class Calliope2017ValidatorAndCollectorVisitor extends CalliopeValidatorAndCollectorVisitor {
+public final class Calliope2017ValidatorAndCollectorVisitor extends MbedValidatorAndCollectorVisitor {
 
     public Calliope2017ValidatorAndCollectorVisitor(ConfigurationAst brickConfiguration, ClassToInstanceMap<IProjectBean.IBuilder> beanBuilders) {
         super(brickConfiguration, beanBuilders);
