+++ RobotMbed/src/main/java/de/fhg/iais/roberta/visitor/validate/CalliopeSimValidatorAndCollectorVisitor.java
@@ -1,26 +1,11 @@
 package de.fhg.iais.roberta.visitor.validate;
 
 import com.google.common.collect.ClassToInstanceMap;
-
 import de.fhg.iais.roberta.bean.IProjectBean;
 import de.fhg.iais.roberta.components.ConfigurationAst;
 import de.fhg.iais.roberta.syntax.action.light.LightAction;
-import de.fhg.iais.roberta.syntax.action.mbed.FourDigitDisplayClearAction;
-import de.fhg.iais.roberta.syntax.action.mbed.FourDigitDisplayShowAction;
-import de.fhg.iais.roberta.syntax.action.mbed.LedBarSetAction;
-import de.fhg.iais.roberta.syntax.action.mbed.MotionKitDualSetAction;
-import de.fhg.iais.roberta.syntax.action.mbed.MotionKitSingleSetAction;
-import de.fhg.iais.roberta.syntax.action.mbed.RadioReceiveAction;
-import de.fhg.iais.roberta.syntax.action.mbed.RadioSendAction;
-import de.fhg.iais.roberta.syntax.action.mbed.RadioSetChannelAction;
-import de.fhg.iais.roberta.syntax.action.mbed.ServoSetAction;
-import de.fhg.iais.roberta.syntax.action.mbed.SwitchLedMatrixAction;
-import de.fhg.iais.roberta.syntax.sensor.generic.AccelerometerSensor;
-import de.fhg.iais.roberta.syntax.sensor.generic.ColorSensor;
-import de.fhg.iais.roberta.syntax.sensor.generic.GyroSensor;
-import de.fhg.iais.roberta.syntax.sensor.generic.HumiditySensor;
-import de.fhg.iais.roberta.syntax.sensor.generic.PinGetValueSensor;
-import de.fhg.iais.roberta.syntax.sensor.generic.UltrasonicSensor;
+import de.fhg.iais.roberta.syntax.action.mbed.*;
+import de.fhg.iais.roberta.syntax.sensor.generic.*;
 import de.fhg.iais.roberta.syntax.sensor.mbed.RadioRssiSensor;
 import de.fhg.iais.roberta.util.syntax.SC;
 import de.fhg.iais.roberta.visitor.IMbedVisitorWithoutDefault;
@@ -33,7 +18,7 @@ public final class CalliopeSimValidatorAndCollectorVisitor extends MbedValidator
 
     @Override
     public Void visitUltrasonicSensor(UltrasonicSensor ultrasonicSensor) {
-        addWarningToPhrase(ultrasonicSensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(ultrasonicSensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitUltrasonicSensor(ultrasonicSensor);
     }
 
@@ -51,21 +36,21 @@ public final class CalliopeSimValidatorAndCollectorVisitor extends MbedValidator
 
     @Override
     public Void visitRadioReceiveAction(RadioReceiveAction radioReceiveAction) {
-        addWarningToPhrase(radioReceiveAction, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(radioReceiveAction, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitRadioReceiveAction(radioReceiveAction);
     }
 
     @Override
     public Void visitPinGetValueSensor(PinGetValueSensor pinValueSensor) {
         if ( pinValueSensor.getMode().equals(SC.PULSEHIGH) || pinValueSensor.getMode().equals(SC.PULSELOW) || pinValueSensor.getMode().equals(SC.PULSE) ) {
-            addWarningToPhrase(pinValueSensor, "SIM_BLOCK_NOT_SUPPORTED");
+            addErrorToPhrase(pinValueSensor, "SIM_BLOCK_NOT_SUPPORTED");
         }
         return super.visitPinGetValueSensor(pinValueSensor);
     }
 
     @Override
     public Void visitColorSensor(ColorSensor colorSensor) {
-        addWarningToPhrase(colorSensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(colorSensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitColorSensor(colorSensor);
     }
 
@@ -77,19 +62,19 @@ public final class CalliopeSimValidatorAndCollectorVisitor extends MbedValidator
 
     @Override
     public Void visitRadioRssiSensor(RadioRssiSensor radioRssiSensor) {
-        addWarningToPhrase(radioRssiSensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(radioRssiSensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitRadioRssiSensor(radioRssiSensor);
     }
 
     @Override
     public Void visitAccelerometerSensor(AccelerometerSensor accelerometerSensor) {
-        addWarningToPhrase(accelerometerSensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(accelerometerSensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitAccelerometerSensor(accelerometerSensor);
     }
 
     @Override
     public Void visitGyroSensor(GyroSensor gyroSensor) {
-        addWarningToPhrase(gyroSensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(gyroSensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitGyroSensor(gyroSensor);
     }
 
@@ -113,7 +98,7 @@ public final class CalliopeSimValidatorAndCollectorVisitor extends MbedValidator
 
     @Override
     public Void visitHumiditySensor(HumiditySensor humiditySensor) {
-        addWarningToPhrase(humiditySensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(humiditySensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitHumiditySensor(humiditySensor);
     }
