--- RobotMbed/src/main/java/de/fhg/iais/roberta/visitor/validate/CalliopeValidatorAndCollectorVisitor.java
@@ -1,13 +1,0 @@
-package de.fhg.iais.roberta.visitor.validate;
-
-import com.google.common.collect.ClassToInstanceMap;
-
-import de.fhg.iais.roberta.bean.IProjectBean;
-import de.fhg.iais.roberta.components.ConfigurationAst;
-
-public class CalliopeValidatorAndCollectorVisitor extends MbedValidatorAndCollectorVisitor {
-
-    public CalliopeValidatorAndCollectorVisitor(ConfigurationAst brickConfiguration, ClassToInstanceMap<IProjectBean.IBuilder> beanBuilders) {
-        super(brickConfiguration, beanBuilders);
-    }
-}
