+++ RobotMbed/src/main/java/de/fhg/iais/roberta/visitor/validate/MicrobitSimValidatorAndCollectorVisitor.java
@@ -1,18 +1,9 @@
 package de.fhg.iais.roberta.visitor.validate;
 
 import com.google.common.collect.ClassToInstanceMap;
-
 import de.fhg.iais.roberta.bean.IProjectBean;
 import de.fhg.iais.roberta.components.ConfigurationAst;
-import de.fhg.iais.roberta.syntax.action.mbed.BothMotorsOnAction;
-import de.fhg.iais.roberta.syntax.action.mbed.BothMotorsStopAction;
-import de.fhg.iais.roberta.syntax.action.mbed.FourDigitDisplayClearAction;
-import de.fhg.iais.roberta.syntax.action.mbed.FourDigitDisplayShowAction;
-import de.fhg.iais.roberta.syntax.action.mbed.LedBarSetAction;
-import de.fhg.iais.roberta.syntax.action.mbed.RadioReceiveAction;
-import de.fhg.iais.roberta.syntax.action.mbed.RadioSendAction;
-import de.fhg.iais.roberta.syntax.action.mbed.RadioSetChannelAction;
-import de.fhg.iais.roberta.syntax.action.mbed.SwitchLedMatrixAction;
+import de.fhg.iais.roberta.syntax.action.mbed.*;
 import de.fhg.iais.roberta.syntax.action.motor.MotorOnAction;
 import de.fhg.iais.roberta.syntax.sensor.generic.AccelerometerSensor;
 import de.fhg.iais.roberta.syntax.sensor.generic.HumiditySensor;
@@ -42,14 +33,14 @@ public final class MicrobitSimValidatorAndCollectorVisitor extends MbedValidator
 
     @Override
     public Void visitRadioReceiveAction(RadioReceiveAction radioReceiveAction) {
-        addWarningToPhrase(radioReceiveAction, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(radioReceiveAction, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitRadioReceiveAction(radioReceiveAction);
     }
 
     @Override
     public Void visitPinGetValueSensor(PinGetValueSensor pinValueSensor) {
         if ( pinValueSensor.getMode().equals(SC.PULSEHIGH) || pinValueSensor.getMode().equals(SC.PULSELOW) || pinValueSensor.getMode().equals(SC.PULSE) ) {
-            addWarningToPhrase(pinValueSensor, "SIM_BLOCK_NOT_SUPPORTED");
+            addErrorToPhrase(pinValueSensor, "SIM_BLOCK_NOT_SUPPORTED");
         }
         return super.visitPinGetValueSensor(pinValueSensor);
     }
@@ -62,13 +53,13 @@ public final class MicrobitSimValidatorAndCollectorVisitor extends MbedValidator
 
     @Override
     public Void visitRadioRssiSensor(RadioRssiSensor radioRssiSensor) {
-        addWarningToPhrase(radioRssiSensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(radioRssiSensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitRadioRssiSensor(radioRssiSensor);
     }
 
     @Override
     public Void visitAccelerometerSensor(AccelerometerSensor accelerometerSensor) {
-        addWarningToPhrase(accelerometerSensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(accelerometerSensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitAccelerometerSensor(accelerometerSensor);
     }
 
@@ -110,13 +101,13 @@ public final class MicrobitSimValidatorAndCollectorVisitor extends MbedValidator
 
     @Override
     public Void visitHumiditySensor(HumiditySensor humiditySensor) {
-        addWarningToPhrase(humiditySensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(humiditySensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitHumiditySensor(humiditySensor);
     }
 
     @Override
     public Void visitUltrasonicSensor(UltrasonicSensor ultrasonicSensor) {
-        addWarningToPhrase(ultrasonicSensor, "SIM_BLOCK_NOT_SUPPORTED");
+        addErrorToPhrase(ultrasonicSensor, "SIM_BLOCK_NOT_SUPPORTED");
         return super.visitUltrasonicSensor(ultrasonicSensor);
     }
