+++ RobotMbed/src/main/java/de/fhg/iais/roberta/worker/codegen/MbedStackMachineGeneratorWorker.java
@@ -1,9 +1,10 @@
-package de.fhg.iais.roberta.worker;
+package de.fhg.iais.roberta.worker.codegen;
 
 import de.fhg.iais.roberta.bean.UsedHardwareBean;
 import de.fhg.iais.roberta.components.Project;
 import de.fhg.iais.roberta.visitor.codegen.MbedStackMachineVisitor;
 import de.fhg.iais.roberta.visitor.lang.codegen.AbstractStackMachineVisitor;
+import de.fhg.iais.roberta.worker.AbstractStackMachineGeneratorWorker;
 
 public class MbedStackMachineGeneratorWorker extends AbstractStackMachineGeneratorWorker {
     @Override
