+++ RobotMbed/src/main/java/de/fhg/iais/roberta/worker/codegen/MicrobitPythonGeneratorWorker.java
@@ -1,11 +1,11 @@
-package de.fhg.iais.roberta.worker;
+package de.fhg.iais.roberta.worker.codegen;
 
 import com.google.common.collect.ClassToInstanceMap;
-
 import de.fhg.iais.roberta.bean.IProjectBean;
 import de.fhg.iais.roberta.components.Project;
 import de.fhg.iais.roberta.visitor.codegen.MicrobitPythonVisitor;
 import de.fhg.iais.roberta.visitor.lang.codegen.AbstractLanguageVisitor;
+import de.fhg.iais.roberta.worker.AbstractLanguageGeneratorWorker;
 
 public class MicrobitPythonGeneratorWorker extends AbstractLanguageGeneratorWorker {
