+++ RobotMbed/src/main/java/de/fhg/iais/roberta/worker/validate/MbedConfigurationValidatorWorker.java
@@ -1,13 +1,4 @@
-package de.fhg.iais.roberta.worker;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+package de.fhg.iais.roberta.worker.validate;
 
 import de.fhg.iais.roberta.components.Project;
 import de.fhg.iais.roberta.syntax.configuration.ConfigurationComponent;
@@ -15,6 +6,10 @@ import de.fhg.iais.roberta.typecheck.NepoInfo;
 import de.fhg.iais.roberta.util.Key;
 import de.fhg.iais.roberta.util.dbc.DbcException;
 import de.fhg.iais.roberta.util.syntax.SC;
+
+import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class MbedConfigurationValidatorWorker {
     private static final String PORT = "PIN1";
