+++ RobotMbed/src/main/java/de/fhg/iais/roberta/worker/validate/MicrobitSimValidatorAndCollectorWorker.java
@@ -1,13 +1,12 @@
-package de.fhg.iais.roberta.worker;
+package de.fhg.iais.roberta.worker.validate;
 
 import com.google.common.collect.ClassToInstanceMap;
-
 import de.fhg.iais.roberta.bean.IProjectBean;
 import de.fhg.iais.roberta.components.Project;
 import de.fhg.iais.roberta.visitor.validate.CommonNepoValidatorAndCollectorVisitor;
 import de.fhg.iais.roberta.visitor.validate.MicrobitSimValidatorAndCollectorVisitor;
 
-public class MicrobitSimValidatorAndCollectorWorker extends AbstractValidatorAndCollectorWorker {
+public class MicrobitSimValidatorAndCollectorWorker extends MicrobitValidatorAndCollectorWorker {
 
     @Override
     protected CommonNepoValidatorAndCollectorVisitor getVisitor(Project project, ClassToInstanceMap<IProjectBean.IBuilder> beanBuilders) {
