+++ RobotMbed/src/main/java/de/fhg/iais/roberta/worker/validate/MicrobitValidatorAndCollectorWorker.java
@@ -1,17 +1,16 @@
-package de.fhg.iais.roberta.worker;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+package de.fhg.iais.roberta.worker.validate;
 
 import com.google.common.collect.ClassToInstanceMap;
-
 import de.fhg.iais.roberta.bean.IProjectBean;
 import de.fhg.iais.roberta.components.Project;
 import de.fhg.iais.roberta.util.syntax.SC;
 import de.fhg.iais.roberta.visitor.validate.CommonNepoValidatorAndCollectorVisitor;
 import de.fhg.iais.roberta.visitor.validate.MicrobitValidatorAndCollectorVisitor;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class MicrobitValidatorAndCollectorWorker extends MbedValidatorAndCollectorWorker {
     public static final List<String> FREE_PINS = Stream.of("0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "19", "20").collect(Collectors.toList());
