+++ RobotMbed/src/main/resources/calliope2017NoBlue.properties
@@ -22,28 +22,28 @@ robot.info.de = https://calliope.cc/
 robot.info.en = https://calliope.cc/en/
 robot.configuration = true
 # either new or old-<sensorprefix> with <sensorprefix>="S" for all robots but mbot, which needs "P"
-robot.configuration.type = new
+robot.configuration.type=new
 # still necessary for backwards compatibility with old configurations
-robot.configuration.old.toplevelblock = mbedBrick_Calliope-Brick
+robot.configuration.old.toplevelblock=mbedBrick_Calliope-Brick
 
 # flag that compiled programs are send back to the client
-robot.connection = autoConnection
+robot.connection=autoConnection
 
-robot.helperMethods = classpath:/calliope.methods.yml
+robot.helperMethods=classpath:/calliope.methods.yml
 
-robot.plugin.worker.validate.and.collect = de.fhg.iais.roberta.worker.Calliope2017NoBlueValidatorAndCollectorWorker
-robot.plugin.worker.validate.and.collect.sim = de.fhg.iais.roberta.worker.CalliopeSimValidatorAndCollectorWorker
-robot.plugin.worker.generate = de.fhg.iais.roberta.worker.CalliopeCxxGeneratorWorker
-robot.plugin.worker.setup = de.fhg.iais.roberta.worker.CompilerSetupWorker
-robot.plugin.worker.compile = de.fhg.iais.roberta.worker.CalliopeCompilerWorker
-robot.plugin.worker.generatesimulation = de.fhg.iais.roberta.worker.MbedStackMachineGeneratorWorker
-robot.plugin.worker.resetFirmware = de.fhg.iais.roberta.worker.MbedResetFirmwareWorker
-robot.plugin.worker.transform.two2three = de.fhg.iais.roberta.worker.MbedTwo2ThreeTransformerWorker
-robot.plugin.worker.transform.three2threeone = de.fhg.iais.roberta.worker.MbedThree2ThreeOneTransformerWorker
+robot.plugin.worker.validate.and.collect=de.fhg.iais.roberta.worker.validate.CalliopeValidatorAndCollectorWorker
+robot.plugin.worker.validate.and.collect.sim=de.fhg.iais.roberta.worker.validate.CalliopeSimValidatorAndCollectorWorker
+robot.plugin.worker.generate=de.fhg.iais.roberta.worker.codegen.CalliopeCxxGeneratorWorker
+robot.plugin.worker.setup=de.fhg.iais.roberta.worker.CompilerSetupWorker
+robot.plugin.worker.compile=de.fhg.iais.roberta.worker.CalliopeCompilerWorker
+robot.plugin.worker.generatesimulation=de.fhg.iais.roberta.worker.codegen.MbedStackMachineGeneratorWorker
+robot.plugin.worker.resetFirmware=de.fhg.iais.roberta.worker.MbedResetFirmwareWorker
+robot.plugin.worker.transform.two2three=de.fhg.iais.roberta.worker.MbedTwo2ThreeTransformerWorker
+robot.plugin.worker.transform.three2threeone=de.fhg.iais.roberta.worker.MbedThree2ThreeOneTransformerWorker
 
-robot.plugin.workflow.showsource = validate.and.collect,generate
-robot.plugin.workflow.compile = validate.and.collect,generate,setup,compile
-robot.plugin.workflow.run = validate.and.collect,generate,setup,compile
+robot.plugin.workflow.showsource=validate.and.collect,generate
+robot.plugin.workflow.compile=validate.and.collect,generate,setup,compile
+robot.plugin.workflow.run=validate.and.collect,generate,setup,compile
 robot.plugin.workflow.getsimulationcode = validate.and.collect.sim,generatesimulation
 robot.plugin.workflow.runnative = setup,compile
 robot.plugin.workflow.compilenative = setup,compile
